pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    { name = "昊色居士", email = "xhrtxh@gmail.com" },
]
dependencies = [
    "qasync>=0.26.0",
    "nuitka>=1.8.4",
    "Pillow>=10.1.0",
    "PySide6-Fluent-Widgets @ git+https://github.com/x-haose/PySide6-Fluent-Widgets.git@PySide6",
    "loguru>=0.7.2",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = { text = "MIT" }

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.0.4",
    "types-ujson>=5.7.0.0",
    "types-PyYAML>=6.0.12.2"
]

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
verify_ssl = true

[tool.pdm.scripts]
check = "pdm run pre-commit run --all-files"
build = "pdm run pyside6-deploy -c pysidedeploy.spec"
